Rewrite simple_to_do example
