--- examples/motoko_examples/simple-to-do/Cargo.toml
@@ -1,9 +1,0 @@
-[workspace]
-members = [
-    "canisters/rust",
-]
-
-[profile.release]
-lto = true
-opt-level = 'z'
-codegen-units = 1
