--- examples/motoko_examples/simple-to-do/benchmark/benchmark.ts
@@ -1,63 +1,0 @@
-import { Benchmark, run_benchmarks } from 'azle/benchmark';
-import { createActor as createActorAzle } from '../dfx_generated/azle';
-import { createActor as createActorMotoko } from '../dfx_generated/motoko';
-import { createActor as createActorRust } from '../dfx_generated/rust';
-
-const short_todo = 'Perform benchmark tests';
-const medium_todo = `Add a few more to dos to this list and make sure that we
-    have pretty long tasks in this test so we can see how having a longer
-    string would affect our benchmarks, it's important to make sure we are
-    testing bounds but also being realistic`;
-const long_todo = `Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
-    do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad
-    minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex
-    ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate
-    velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat
-    cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id
-    est laborum. Sed ut perspiciatis unde omnis iste natus error sit voluptatem
-    accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab
-    illo inventore veritatis et quasi architecto beatae vitae dicta sunt
-    explicabo. Nemo enim ipsam voluptatem quia voluptas sit aspernatur aut odit
-    aut fugit, sed quia consequuntur magni dolores eos qui ratione voluptatem
-    sequi nesciunt. Neque porro quisquam est, qui dolorem ipsum quia dolor sit
-    amet, consectetur, adipisci velit, sed quia non numquam eius modi tempora
-    incidunt ut labore et dolore magnam aliquam quaerat voluptatem. Ut enim ad
-    minima veniam, quis nostrum exercitationem ullam corporis suscipit
-    laboriosam, nisi ut aliquid ex ea commodi consequatur? Quis autem vel eum
-    iure reprehenderit qui in ea voluptate velit esse quam nihil molestiae
-    consequatur, vel illum qui dolorem eum fugiat quo voluptas nulla pariatur?`;
-
-const benchmarks: Benchmark[] = [
-    {
-        canister_method: 'add_todo',
-        benchmark_description: 'short todo description',
-        args: [short_todo]
-    },
-    {
-        canister_method: 'add_todo',
-        benchmark_description: 'medium todo description',
-        args: [medium_todo]
-    },
-    {
-        canister_method: 'add_todo',
-        benchmark_description: 'long todo description',
-        args: [long_todo]
-    },
-    {
-        canister_method: 'complete_todo',
-        args: [1n]
-    },
-    {
-        canister_method: 'clear_completed'
-    }
-];
-
-run_benchmarks(
-    'GLOBAL_VALUE',
-    benchmarks,
-    createActorAzle,
-    createActorMotoko,
-    createActorRust,
-    10,
-    `benchmarks`
-);
