--- examples/motoko_examples/simple-to-do/benchmarks.csv
@@ -1,6 +1,0 @@
-Description,Azle Wasm Instructions (function body only),Azle Wasm Instructions (function prelude and body),Motoko Wasm Instructions (function body only),Motoko Wasm Instructions (function prelude and body),Rust Wasm Instructions (function body only),Rust Wasm Instructions (function prelude and body),Azle/Motoko Change Multiplier (function body only),Azle/Motoko Change Multiplier (function prelude and body),Azle/Rust Change Multiplier (function body only),Azle/Rust Change Multiplier (function prelude and body),Motoko/Azle Change Multiplier (function body only),Motoko/Azle Change Multiplier (function prelude and body),Motoko/Rust Change Multiplier (function body only),Motoko/Rust Change Multiplier (function prelude and body),Rust/Azle Change Multiplier (function body only),Rust/Azle Change Multiplier (function prelude and body),Rust/Motoko Change Multiplier (function body only),Rust/Motoko Change Multiplier (function prelude and body)
-add_todo: short todo description,306394,876000,3662,5659,3589,40538,125,185,101,22,-125,-185,-0,-9,-101,-22,0,9
-add_todo: medium todo description,307713,1099141,4749,8952,3731,41826,112,154,110,27,-112,-154,-0,-6,-110,-27,0,6
-add_todo: long todo description,311881,903009,2593,15956,4860,52978,120,57,107,17,-120,-57,-1,-3,-107,-17,1,3
-complete_todo,381408,958996,3009,4485,2645,42565,127,214,144,23,-127,-214,1,-9,-144,-23,-1,9
-clear_completed,10920295,11458636,71115,72246,197058,216157,154,159,55,53,-154,-159,-3,-3,-55,-53,3,3
