--- examples/motoko_examples/simple-to-do/canisters/azle/index.did
@@ -1,10 +1,0 @@
-type PerfResult = record { wasmBodyOnly : nat64; wasmIncludingPrelude : nat64 };
-type ToDo = record { completed : bool; description : text };
-service : () -> {
-  addTodo : (text) -> (nat);
-  clearCompleted : () -> ();
-  completeTodo : (nat) -> ();
-  getPerfResult : () -> (opt PerfResult) query;
-  getTodos : () -> (vec ToDo) query;
-  showTodos : () -> (text) query;
-}
