--- examples/motoko_examples/simple-to-do/canisters/rust/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "rust"
-version = "0.0.0"
-edition = "2018"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-ic-cdk = "0.5.2"
-ic-cdk-macros = "0.5.2"
-candid = "0.7.14"
