--- examples/motoko_examples/simple-to-do/canisters/rust/src/lib.did
@@ -1,19 +1,0 @@
-type ToDo = record {
-    "id": nat;
-    "description": text;
-    "completed": bool;
-};
-
-type PerfResult = record {
-    "wasm_body_only": nat64;
-    "wasm_including_prelude": nat64;
-};
-
-service: {
-    "get_todos": () -> (vec ToDo) query;
-    "show_todos": () -> (text) query;
-    "add_todo": (text) -> (nat);
-    "complete_todo": (nat) -> ();
-    "clear_completed": () -> ();
-    "get_perf_result": () -> (opt PerfResult) query;
-}
