+++ examples/motoko_examples/simple-to-do/dfx.json
@@ -1,31 +1,14 @@
 {
     "canisters": {
-        "azle": {
+        "simple_to_do": {
             "type": "custom",
-            "build": "npx azle azle",
-            "root": "canisters/azle",
-            "ts": "canisters/azle/index.ts",
-            "candid": "canisters/azle/index.did",
-            "wasm": ".azle/azle/azle.wasm.gz",
+            "build": "npx azle simple_to_do",
+            "root": "src",
+            "ts": "src/index.ts",
+            "candid": "src/index.did",
+            "wasm": ".azle/simple_to_do/simple_to_do.wasm.gz",
             "declarations": {
-                "output": "dfx_generated/azle",
-                "node_compatibility": true
-            }
-        },
-        "motoko": {
-            "type": "motoko",
-            "main": "canisters/motoko/main.mo",
-            "declarations": {
-                "output": "dfx_generated/motoko",
-                "node_compatibility": true
-            }
-        },
-        "rust": {
-            "type": "rust",
-            "package": "rust",
-            "candid": "canisters/rust/src/lib.did",
-            "declarations": {
-                "output": "dfx_generated/rust",
+                "output": "test/dfx_generated/simple_to_do",
                 "node_compatibility": true
             }
         }
