+++ examples/motoko_examples/simple-to-do/package.json
@@ -8,8 +8,8 @@
         "azle": "0.17.1"
     },
     "devDependencies": {
-        "@dfinity/agent": "^0.12.1",
-        "ts-node": "10.9.1",
-        "typescript": "^4.7.4"
+        "@dfinity/agent": "^0.19.2",
+        "ts-node": "^10.9.1",
+        "typescript": "^5.2.2"
     }
 }
