+++ examples/motoko_examples/simple-to-do/test/pretest.ts
@@ -3,15 +3,15 @@ import { execSync } from 'child_process';
 async function pretest() {
     await new Promise((resolve) => setTimeout(resolve, 5000));
 
-    execSync(`dfx canister uninstall-code azle || true`, {
+    execSync(`dfx canister uninstall-code simple_to_do || true`, {
         stdio: 'inherit'
     });
 
-    execSync(`dfx deploy azle`, {
+    execSync(`dfx deploy simple_to_do`, {
         stdio: 'inherit'
     });
 
-    execSync(`dfx generate azle`, {
+    execSync(`dfx generate simple_to_do`, {
         stdio: 'inherit'
     });
 }
