+++ examples/motoko_examples/simple-to-do/test/test.ts
@@ -1,8 +1,8 @@
 import { getCanisterId, runTests } from 'azle/test';
-import { createActor } from '../dfx_generated/azle/';
+import { createActor } from './dfx_generated/simple_to_do/';
 import { getTests } from './tests';
 
-const todoCanister = createActor(getCanisterId('azle'), {
+const todoCanister = createActor(getCanisterId('simple_to_do'), {
     agentOptions: {
         host: 'http://127.0.0.1:8000'
     }
